scripts/create_pr_comment/requirements.txt
PyGithub
markdown
Jinja2